+++ apps/interface/src/lib/polyfills.ts
@@ -1,5 +1,2 @@
-import { Buffer } from "buffer";
-
-globalThis.Buffer ??= Buffer;
 // @ts-expect-error
-globalThis.process ??= { env: {} };
+globalThis.self ??= globalThis;