+++ packages/sdk/src/zklogin.ts
@@ -1,2 +1,3 @@
-export { ZkLogin } from "./ZkLoginSdk.js";
+export { GoogleProvider } from "./providers/Google.js";
 export { PublicKeyRegistry } from "./PublicKeyRegistry.js";
+export { ZkLogin } from "./ZkLoginSdk.js";